requirements.txt
faiss-cpu
numpy
torch
sentence-transformers
PyPDF2
python-docx
python-magic
tqdm
python-dateutil